+++ jabgui/src/main/java/org/jabref/gui/libraryproperties/saving/SavingPropertiesViewModel.java
@@ -109,9 +109,7 @@ public class SavingPropertiesViewModel implements PropertiesTabViewModel {
     }
 
     @Override
-    public void storeSettings() {
-        MetaData metaData = databaseContext.getMetaData();
-
+    public void storeSettings(MetaData metaData) {
         if (libraryProtectedProperty.getValue()) {
             metaData.markAsProtected();
         } else {